mv read to lib
