+++ src/lib/index.js
@@ -1,5 +1,7 @@
 import tableRule from './rules/table'
 import titleRule from './rules/method-title'
+import { createReadStream } from 'fs'
+import Catchment from 'catchment'
 
 export const getLink = (title) => {
   const l = title
@@ -20,3 +22,14 @@ export const makeARegexFromRule = (rule) => {
 
 export const exactTable = makeARegexFromRule(tableRule)
 export const exactMethodTitle = makeARegexFromRule(titleRule)
+
+export const read = async (source) => {
+  const rs = createReadStream(source)
+  const data = await new Promise(async (r, j) => {
+    const { promise } = new Catchment({ rs })
+    rs.on('error', j)
+    const res = await promise
+    r(res)
+  })
+  return data
+}
